extract async operation inventory route
